refactor: Move UI components to closer location
In this commit:
- Moved several components closer to implementation
- Renamed components/ui to compoenents/base for improved organazation and readability.

This change aims to improve code organization by relocating the components from the '@/components' directory to a more contextually relevant location within 'app/'. This change streamlines the project structure, making it easier to find and maintain UI components where they are primarily used.
