--- components/ui/skeleton.tsx
@@ -1,15 +1,0 @@
-import { cn } from "@/lib/client/utils";
-
-function Skeleton({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLDivElement>) {
-  return (
-    <div
-      className={cn("animate-pulse rounded-md bg-primary/10", className)}
-      {...props}
-    />
-  );
-}
-
-export { Skeleton };
